backend/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
dataclasses-json==0.5.14
exceptiongroup==1.1.3
fastapi==0.103.1
filelock==3.12.3
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
InstructorEmbedding==1.0.1
Jinja2==3.1.2
langchain==0.0.279
langsmith==0.0.33
lit==16.0.6
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.28.0
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyngrok==6.0.0
pypdf==3.15.4
python-decouple==3.8
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
twilio==8.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
yarl==1.9.2
